+++ DrawEngine.Renderer/PhotonMapping/Photon.cs
@@ -26,7 +26,7 @@ namespace DrawEngine.Renderer.PhotonMapping
         public byte Theta; // incoming direction
         //Only for acceleration in search (avoid linear search)
         public int Index = -1;
-        public float PrevRefractIndex = 0;
+        public double PrevRefractIndex = 0;
         public Photon() {}
         public Photon(Vector3D dir, Point3D pos, RGBColor power)
         {
